pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncio-mqtt>=0.16.2",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
]
